Wire up basic message-passing from JupyterLab to WWT
This has an extremely lame security setup right now that makes us
available to XSS attacks. If/when the WWT client app starts having some
sense of a "login" such that it can perform actions in the user's name,
we'll need to figure out a setup so that it can't be abused.