Check and report availability of the Kernel Data Relay
I'm creating a new Jupyter Server extension, `wwt_kernel_data_relay`,
that will improve the way that pywwt can publish data to the Jupyter
Server. This work is necessary to set us up to tile large imagery on the
pywwt side and actually make such imagery accessible inside the Jupyter
environment -- the previous approach wasn't scalable to lots of files,
and it didn't work in the case when the Python kernel and Jupyter server
run on different machines.

For a good overall UX, we need to have some insight as to whether the
desired server extension is actually available. Here we extend the
JupyerLab extension to probe for the availability of the KDR, and to
report its status to kernels that try to connect to the app. It's a very
simple mechanism but should do what we need. It feels a bit circuitous
to have the kernel learn about the availability of the server extension
by talking to the frontend, but I think that's how things actually have
to work. (I think it would be a lot more fragile to put in something to
have the kernel get explicit confirmation from the server; here we're
building on the guaranteed comms infrastructure.)