src/viewer.ts: relay all messages with the appropriate origin
We were only relaying view-state messages, but this is actually quite
wrong: there are going to be lots of bidirectional communications
between the app and clients, such as requests for data or notifications
that requests have completed. The JupyterLab extension shouldn't be in
the business of having to know what all of these messages are. Just
forward along everything so long as the origin is OK.