Add _pywwtExpedite tags for pywwt's Jupyter async skullduggery
In order for user code to be able to interact with WWT with proper async
characteristics, we need to hack the Python kernel to expedite
processing of WWT messages, for reasons laid out in detail in the
corresponding pywwt code. Here we add the special tag that requests
expedited processing. If the kernel isn't hacked, then nothing unusual
happens.