Support relative and baseUrl-relative app URLs
Because we constructed a URL object from the appUrl setting, it had to
be an absolute URL. Here we emulate the logic from pywwt to prepend the
JupyterLab baseUrl if the URL starts with `/wwt`, and also to make it an
absolute URL if required. This allows us to successfully work with
user-specific appUrls, such as the local version of the research app
provided by pywwt.